requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
scikit-learn>=1.1.0
yfinance>=0.2.0
arch>=6.2.0
statsmodels>=0.14.0
torch>=2.0.0
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=1.10.0
redis>=4.5.0
mlflow>=2.2.0
ta>=0.10.2
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.8.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.13.0
pytest>=7.2.0
PyYAML>=6.0
python-dotenv>=1.0.0
joblib>=1.2.0
hmmlearn>=0.3.0
